Add rfc command and revise data receiving
The 'rfc' command sends the two IRC rfc's to the channel. This may be
useful in developing an IRC bot.

Instead of using socket.recv(), a file can be created from the socket
using socket.makefile and call its readline method. The file is created
with a binary mode so we can decode manually later.

To prevent the bot from crashing because of UnicodeDecodeErrors, the
str.decode method is passed errors='ignore'.

 plugins/standard.py |  5 +++++
 utils/irc.py        | 15 ++++++---------
 2 files changed, 11 insertions(+), 9 deletions(-)