Add register_callback(s) functions
The register_callbacks function automatically searches for callbacks in
the bot object, and registers these. If a plugin wants to register a
callback, they can use register_callback(callback_type, function).

This commit implements these functions, and calls register_callbacks on
bot startup. This is part two of the Infobot __IRCCallbacks__ system.