Replace try_to_call with run_callback
Phew. This might be my favourite change I've ever made to James - the
try_to_call function was ugly, had a bazillion options that were never
used, and the way of using it in the main loop was just awful.

run_callback is the last piece of the Infobot __irccallbacks__ puzzle -
it is used to call callbacks in the main loop. Because IRCCallback can
take numeric method names, the three(!) calls to try_to_call can be
replaced with a single call to run_callback.